♻️refactor: MVC 패턴에서 아토믹 패턴으로 변경 - Controller에서 담당하던 날씨 관련 API MyLocationWeather 컴포넌트 내부로 이동
